Enum goto::commands::Cmd

pub enum Cmd {
    Get(ShortPath),
    Add(ShortPath),
    Edit(ShortPath),
    Rm(String),
    Del(String),
    Decr(u32),
    Reset,
}

enum to store the command to execute, and its arguments

Examples

let cmd = Cmd::Get([Some(String::from("short")), Some(String::from("path"))]);

Note

The command is used to get the path of a directory, to add a directory to the file of supported shortcuts, to edit the path of a directory, to remove a directory from the file of supported shortcuts, to delete a directory from the file of supported shortcuts, to decrement the usage of all directories, or to reset the usage of all directories to 0.

Variants

Get(ShortPath)

Get the path of a directory.

Add(ShortPath)

Add a directory to the file of supported shortcuts.

Edit(ShortPath)

Edit the path of a directory.

Rm(String)

Remove a directory from the file of supported shortcuts.

Del(String)

Delete a directory from the file of supported shortcuts.

Decr(u32)

Decrement the usage of all directories.

Reset

Reset the usage of all directories to 0.

Implementations

impl Cmd

pub fn append(&mut self, value: String)

Implement the append method for the Cmd enum.

Arguments

value - The value to append to the command.

Panics

If the command is already full of all its arguments.

Examples

let mut cmd = Cmd::Get([None, None]);
cmd.append(String::from("short"));
cmd.append(String::from("path"));