Module goto::dirs

This module contains the functions to push and pop directories from the history file.

Structs

DirsLine 🔒

Structure to contain the data of a line in the directory file

SearchState 🔒

Structure to contain the state of the search

Functions

add 🔒

Function to add a shortcut to a line of the directory file

edit 🔒

Arguments

get 🔒

Function to get the path of a directory from the directory file

read

Function to read the directory file

read_dline 🔒

Function to read a line of the directory file

remove 🔒

Function to remove a shortcut from a line of the directory file

state

Function to print state of the directories

std_path 🔒