Function goto::dirs::read_dline

fn read_dline(
    rdline: &str,
    args: &[Cmd],
    success: &mut bool,
    incr: u32,
    sstate: &mut SearchState
) -> String

Function to read a line of the directory file

Arguments

rdline - The line of the directory file

args - The arguments of the command

success - A mutable reference to a boolean to indicate if the command was successful

incr - The increment value

sstate - A mutable reference to the state of the search

Returns

The line of the directory file

Example

let mut success = false;
let mut sstate = SearchState::default();
let rdline = "/home/user/folder;f;folder;1";
let args = vec!["get".to_string(), "f".to_string()];
let line = read_dline(&rdline, &args, &mut success, 1, &mut sstate);

Panics

If the command is invalid