Function goto::vscode

pub(crate) fn vscode(args2: &[String], path: &str)

Open the directory in Visual Studio Code.

Arguments

args2 - The arguments of the command

path - The path of the directory to open

Note

This function is used to open the directory in Visual Studio Code, if the code argument is present. The function uses the code command to open the directory in Visual Studio Code. The function raises a warning if the code command is not found. The function is called after finding the path of the directory to go to, and after updating the usage of the directory.