Function goto::hist::pushd

pub fn pushd(histpath: &str, path: &str)

Push a directory to the history file.

Arguments

histpath - The path to the history file

path - The path to the directory to push

Example

pushd("lib/hist.csv", "/home/user/folder");

Note

The history file is a simple text file with the following format:

/home/user/folder1;pid1;time1
/home/user/folder2;pid2;time2

Where pid is the process id of the process that pushed the directory and time is the time in seconds since the Unix Epoch. The last line is the most recent directory pushed.